Add files via upload
This was the result of a 24 hour Hackathon. Collaborating with two peers, we created a visual art game which uses C# and mono compiling. It still requires refinement, but it was able to be demoed at Pearl Hacks 2022. The goal was to create a tetris-style art game which could be used to express creativity.